Move http independent classes to util.paging; Use Response<T> to hide Map.Entry from public functions
