+++ sdk/core/azure-core/src/main/java/com/azure/android/core/util/paging/PagedDataCollection.java
@@ -2,9 +2,7 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-package com.azure.android.core.http.responsepaging;
-
-import com.azure.android.core.util.paging.Page;
+package com.azure.android.core.util.paging;
 
 import java.util.Iterator;
 import java.util.LinkedHashMap;
